+++ persistence/src/main/java/com/furiosaming/walletService/persistence/model/BankAccount.java
@@ -17,7 +17,7 @@ public class BankAccount {
     /** Уникальный идентификатор банковского счета */
     private Long id;
     /** Сумма, которая находится на счету в данный момент */
-    private double cashValue;
+    private long cashValue;
     /** Список транзакций, проводившихся на данном счете */
     private List<Transaction> transactions = new ArrayList<>();
     /** Список действий, связанных с аккаунтом пользователя */
